Fix NaN Bug
This commit when implemented fixes a bug where the calculator screen would return NaN to the user even though no calculation request was made.

This bug originated from the addDecimal function. addDecimal code would execute second code block because function did not check if user input was indeed a decimal. This would occasionally lead to a decimal being added and evaluated when tapping on an operator button.