Add percent and invertSign functions
When implemented this commit will add the percent and invertSign functions to the calculator.

percent function
When the user taps the percent button, the value on the calculator screen will be converted to a percent by dividing the value on the screen by 100.

invertSign function
When the user taps the +/- button on the calculator, any number > 0 will have its sign inverted. A negative number will be converted to a positive number and a negative number will be converted to a positive number.
NOTE: If user taps on an operator, the +/- button will remain inactive until the user begins to input their second number.